REFACTOR: dto 위치 리팩토링 See also: #16
