+++ src/main/java/com/sparta/deliveryi/ai/application/service/AiLogService.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.deliveryi.ai.domain.AiLog;
 import com.sparta.deliveryi.ai.domain.service.AiLogRegister;
 import com.sparta.deliveryi.ai.domain.AiStatus;
-import com.sparta.deliveryi.ai.presentation.AiLogRequest;
-import com.sparta.deliveryi.ai.presentation.AiLogResponse;
+import com.sparta.deliveryi.ai.presentation.dto.AiLogRequest;
+import com.sparta.deliveryi.ai.presentation.dto.AiLogResponse;
 import com.sparta.deliveryi.ai.infrastructure.dto.GeminiRequest;
 import com.sparta.deliveryi.ai.infrastructure.dto.GeminiResponse;
 import lombok.RequiredArgsConstructor;
