+++ src/main/java/com/sparta/deliveryi/ai/presentation/dto/AiLogRequest.java
@@ -1,4 +1,4 @@
-package com.sparta.deliveryi.ai.presentation;
+package com.sparta.deliveryi.ai.presentation.dto;
 
 public record AiLogRequest(
         Long menuId,
